Move require_permission/require_method to senaite.core.decorators
They are generic endpoint decorators, not tied to the base view classes, so
they belong in the decorators package next to readonly_transaction/retriable
rather than in browser/views.py.
